🐱✨
Support float-precision blockstate rotation (22.5° increments)

- BlockstateJson.Variant: int x/y → float x/y
- BakedModelCache: buildKey/BakeRequest/CacheLoader int→float with
  compact float formatting (integer angles without decimal point)
- JsonModelBake: applyYRotation/applyXRotation int→float deg params
- UnbakedModel + ModelJsonUnbakedAdapter: bake interface int→float
- ModelBaker + BakingCore: all bake overloads int→float
- ModelRegistry: bakeModelPart int→float
- AsyncBakePipeline: BakeTask fields + generateTasks int→float
- RenderDispatcher: ResolvedModel.rot int→float
- ResidentStateModel: local rotX/rotY int→float
- RenderSubmit: rotationDeg field + constructors + accessor int→float
- RenderSubmitView: rotationDeg() return type int→float
- UniformRenderPipeline: all renderBlockQuads overloads int→float
- BlockstateKeyValidator: relax validation from 90° to 22.5° multiples
  with float tolerance (±0.01°)

+++ src/main/java/decok/dfcdvadstf/catframe/model/ModelRegistry.java
@@ -64,7 +64,7 @@ public class ModelRegistry {
         /**
          * Public API: bake a model path into a BlockStateModelPart with Y rotation.
          */
-        public static BlockStateModelPart bakeModelPart(String modelPath, int rotationY) {
+        public static BlockStateModelPart bakeModelPart(String modelPath, float rotationY) {
             return bakeModelPart(modelPath, 0, rotationY);
         }
 
@@ -73,7 +73,7 @@ public class ModelRegistry {
          * [W3] 支持 blockstate 中的 x 旋转字段。
          * 通过 {@link BakedModelCache} 懒烘焙，线程安全。
          */
-        public static BlockStateModelPart bakeModelPart(String modelPath, int rotationX, int rotationY) {
+        public static BlockStateModelPart bakeModelPart(String modelPath, float rotationX, float rotationY) {
             String cacheKey = BakedModelCache.buildKey(modelPath, rotationX, rotationY);
             BlockStateModelPart part = BakedModelCache.INSTANCE.get(cacheKey);
             return part != null ? part : BlockStateModelPart.empty();
